Move StudioTree to components
